removed db click functionality
