+++ src/resources/items/ItemList.tsx
@@ -38,7 +38,7 @@ import DestroyRestoreItems from './DestroyRestoreItems'
 import { AuditType } from '../../utils/activity-types'
 import useAudit from '../../hooks/useAudit'
 import BooleanFilter from '../../components/BooleanFilter'
-import DblClickDatagridConfigurable from '../../components/DblClickDatagridConfigurable'
+import DatagridConfigurableWithShow from '../../components/DatagridConfigurableWithShow'
 import { RestoreFromTrash } from '@mui/icons-material'
 import DispatchItems from './DispatchItems'
 import List from '../../components/ListWithLocalStore'
@@ -460,7 +460,7 @@ export default function ItemList(props?: ItemListType): React.ReactElement {
       {typeof children !== 'undefined' ? (
         children
       ) : (
-        <DblClickDatagridConfigurable
+        <DatagridConfigurableWithShow
           resource={constants.R_ITEMS}
           bulkActionButtons={<BulkActions />}
           omit={omitColumns}>
@@ -504,7 +504,7 @@ export default function ItemList(props?: ItemListType): React.ReactElement {
           <DateField source='dispatchedDate' />
           <TextField source='remarks' />
           <TextField source='musterRemarks' />
-        </DblClickDatagridConfigurable>
+        </DatagridConfigurableWithShow>
       )}
     </List>
   )
